macroDjango/requirements.txt
django==2.0
django-redis
djangorestframework
markdown
django-filter
pillow
django-crispy-forms>=1.6.0
django-import-export>=0.5.1
django-reversion>=2.0.0
django-formtools
future==0.15.2
httplib2==0.9.2
six==1.10.0
requests
xlwt
xlsxwriter
coreapi
Pygments
django-guardian
djangorestframework-jwt
pycryptodome
celery
django-celery-results
redis